# api_server/openapi_server/models/car_distance.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model_ import Model
from openapi_server import util


class CarDistance(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, token=None, distance=None, travel_time=None):  # noqa: E501
        """CarDistance - a model defined in OpenAPI

        :param token: The token of this CarDistance.  # noqa: E501
        :type token: str
        :param distance: The distance of this CarDistance.  # noqa: E501
        :type distance: float
        :param travel_time: The travel_time of this CarDistance.  # noqa: E501
        :type travel_time: int
        """
        self.openapi_types = {
            'token': str,
            'distance': float,
            'travel_time': int
        }

        self.attribute_map = {
            'token': 'token',
            'distance': 'distance',
            'travel_time': 'travel_time'
        }

        self._token = token
        self._distance = distance
        self._travel_time = travel_time

    @classmethod
    def from_dict(cls, dikt) -> 'CarDistance':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The CarDistance of this CarDistance.  # noqa: E501
        :rtype: CarDistance
        """
        return util.deserialize_model(dikt, cls)

    @property
    def token(self):
        """Gets the token of this CarDistance.

        the Hyrde car token  # noqa: E501

        :return: The token of this CarDistance.
        :rtype: str
        """
        return self._token

    @token.setter
    def token(self, token):
        """Sets the token of this CarDistance.

        the Hyrde car token  # noqa: E501

        :param token: The token of this CarDistance.
        :type token: str
        """

        self._token = token

    @property
    def distance(self):
        """Gets the distance of this CarDistance.

        distance in km.  # noqa: E501

        :return: The distance of this CarDistance.
        :rtype: float
        """
        return self._distance

    @distance.setter
    def distance(self, distance):
        """Sets the distance of this CarDistance.

        distance in km.  # noqa: E501

        :param distance: The distance of this CarDistance.
        :type distance: float
        """

        self._distance = distance

    @property
    def travel_time(self):
        """Gets the travel_time of this CarDistance.

        Estimated travel time in seconds.  # noqa: E501

        :return: The travel_time of this CarDistance.
        :rtype: int
        """
        return self._travel_time

    @travel_time.setter
    def travel_time(self, travel_time):
        """Sets the travel_time of this CarDistance.

        Estimated travel time in seconds.  # noqa: E501

        :param travel_time: The travel_time of this CarDistance.
        :type travel_time: int
        """

        self._travel_time = travel_time
# flake8: noqa