# api_server/openapi_server/models/car_distances.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from openapi_server.models.base_model_ import Model
from openapi_server.models.car_distance import CarDistance
from openapi_server import util

from openapi_server.models.car_distance import CarDistance  # noqa: E501

class CarDistances(Model):
    """NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).

    Do not edit the class manually.
    """

    def __init__(self, items=None):  # noqa: E501
        """CarDistances - a model defined in OpenAPI

        :param items: The items of this CarDistances.  # noqa: E501
        :type items: List[CarDistance]
        """
        self.openapi_types = {
            'items': List[CarDistance]
        }

        self.attribute_map = {
            'items': 'items'
        }

        self._items = items

    @classmethod
    def from_dict(cls, dikt) -> 'CarDistances':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The CarDistances of this CarDistances.  # noqa: E501
        :rtype: CarDistances
        """
        return util.deserialize_model(dikt, cls)

    @property
    def items(self):
        """Gets the items of this CarDistances.


        :return: The items of this CarDistances.
        :rtype: List[CarDistance]
        """
        return self._items

    @items.setter
    def items(self, items):
        """Sets the items of this CarDistances.


        :param items: The items of this CarDistances.
        :type items: List[CarDistance]
        """

        self._items = items
# flake8: noqa